requirements.txt
opencv-python
mediapipe
numpy
pyalsaaudio
